+++ src/app/trasteos/trasteos.page.ts
@@ -16,9 +16,9 @@ export class TrasteosPage implements OnInit {
   @Input() uid
   @Input() nombre
   @Input() proyecto
+  @Input() apto
 
   
-  apto: number;
   fecha: string
   myDate: Date;
 
@@ -44,26 +44,51 @@ export class TrasteosPage implements OnInit {
     });
 }
 
+async presentAlertBlock(header,text) {
+  const alert = await this.alertController.create({
+    cssClass: 'my-custom-class',
+    header: 'Alerta!',
+    subHeader: header,
+    message: text,
+    buttons: ['OK']
+  });
+
+  await alert.present();
+
+  const { role } = await alert.onDidDismiss();
+  console.log('onDidDismiss resolved with role', role);
+}
+
   upload_publication(){
-    console.log(this.myDate)
-    var ti = moment(this.myDate).format('h:mm:ss a'); 
-     var dt = moment(this.myDate).format('DD-MM-YYYY');
-    let mudanza = {
-        nombre: this.nombre,
-        apto: this.apto,
-        fecha: dt,
-        hora: ti
-        // dia: dt,
-        // hora: ti
-      };
-      var id = Math.floor(Math.random() * 3213546846468435454) + 1
-      console.log("random",id)
-      var sid = id.toString()
-     //this.firestoreService.add("Proyectos/"+this.proyecto+"/mudanzas", mudanza );
-     this.firestoreService.insertar("Proyectos/"+this.proyecto+"/mudanzas/", sid, mudanza )
-     this.firestoreService.insertar("user/"+this.uid+"/proyectos/"+this.proyecto+"/mudanzas/", sid, mudanza )
-     $("#apto").val("")
-     this.presentAlertdone();
+    if (!this.apto) {
+      this.presentAlertBlock("Usuario sin apartamento ","Ve a tu perfil y rellena tu apartamento para poder hacer reservas")
+     } else {
+     
+
+       console.log(this.myDate)
+      var ti = moment(this.myDate).format('h:mm:ss a'); 
+      var dt = moment(this.myDate).format('DD-MM-YYYY');
+      const res = confirm("¿Confirmar trasteo para el dia "+dt+" a las " + ti + "?");
+      if(res){
+            let mudanza = {
+            nombre: this.nombre,
+            apto: this.apto,
+            fecha: dt,
+            hora: ti
+                // dia: dt,
+                // hora: ti
+              };
+              var id = Math.floor(Math.random() * 3213546846468435454) + 1
+              console.log("random",id)
+              var sid = id.toString()
+            //this.firestoreService.add("Proyectos/"+this.proyecto+"/mudanzas", mudanza );
+            this.firestoreService.insertar("Proyectos/"+this.proyecto+"/mudanzas/", sid, mudanza )
+            this.firestoreService.insertar("user/"+this.uid+"/proyectos/"+this.proyecto+"/mudanzas/", sid, mudanza )
+          // $("#apto").val("")
+            this.presentAlertdone();
+      }
+     
+     }
   }
 
   async presentAlert(mensaje) {
@@ -101,20 +126,31 @@ export class TrasteosPage implements OnInit {
   }
 
   delete(comunicado){
-    //console.log("borrando base de datos de",this.current_user_uid,  " del proyecto ",proyecto)
-    this.fbs.delete_doc("Proyectos/"+this.proyecto+"/mudanzas", comunicado).then(() => {
-      // Actualizar la lista completa
-     // this.consultar_lista_servicios();
-      // Limpiar datos de pantalla
-      //this.tareaEditando = {} as Tarea;
-    })
-  
-    // this.fbs.delete_doc("Admins/"+this.current_user_uid+"/proyectos",proyecto).then(() => {
-    //   // Actualizar la lista completa
-    //   this.consultar_lista_servicios();
-    //   // Limpiar datos de pantalla
-    //   //this.tareaEditando = {} as Tarea;
-    // })
+    const res = confirm("Estas seguro que quieres eliminar este trasteo?");
+    if(res){
+      //console.log("borrando base de datos de",this.current_user_uid,  " del proyecto ",proyecto)
+      this.fbs.delete_doc("Proyectos/"+this.proyecto+"/mudanzas", comunicado).then(() => {
+        // Actualizar la lista completa
+      // this.consultar_lista_servicios();
+        // Limpiar datos de pantalla
+        //this.tareaEditando = {} as Tarea;
+      })
+
+      this.fbs.delete_doc("user/"+this.uid+"/proyectos/"+this.proyecto+"/mudanzas", comunicado).then(() => {
+        // Actualizar la lista completa
+      // this.consultar_lista_servicios();
+        // Limpiar datos de pantalla
+        //this.tareaEditando = {} as Tarea;
+      })
+    
+      // this.fbs.delete_doc("Admins/"+this.current_user_uid+"/proyectos",proyecto).then(() => {
+      //   // Actualizar la lista completa
+      //   this.consultar_lista_servicios();
+      //   // Limpiar datos de pantalla
+      //   //this.tareaEditando = {} as Tarea;
+      // })
+    }
+    
     
   }
 }
